Use old buffer write for greyscale.
Previously we used the fact that an update will have the effect, at
the end, of copying the new bitmap to the old bitmap. The next update
would use old/new bits difference to understand if the change it's
WW, BB, WB, BW. However we can set the old framebuffer directly as
per datasheet. The previous system worked, but was kinda of a hack.