Do blocking updates by default.
If I understand correctly reading the very lacking data sheets available
around, the booster may remain turned ON after the display refresh if
not turned off manually via POF command. If that is true, it is safer as
a default update method to block and turn off the display: this may
improve both energy usage and display life.