Remove reaffirming of past color: damages possible.
In old LUTs for WW and BB we used just to force the microcapsules
towards the position they should be. But this for some reasons stresses
the display: there will be some kind of semi-permanent polarization. So
now we keep WW/BB to ground, and from time to time do a full refresh
if we are using a no-flickering mode.